feat: add esm solution
